Finished the Landing Page
Completed all the necessary sections including: the quotation area,
call to action, and the footer. All the sections are aligned correctly
to ensure that it replicates an appealing site. All aesthetics are in
line of what was asked such as the coloring, font, etc. Window can be
minimized without all the elements disappearing from the page.